refactor: extract schema parsing and rendering pipeline
